--- a/tests/test_DRadiosityFast.py
+++ b/tests/test_DRadiosityFast.py
@@ -8,12 +8,12 @@ import sparrowpy as sp
 
 
 def test_init(sample_walls):
-    radiosity = sp.DirectionalRadiosityFast.from_polygon(sample_walls, 0.2)
-    npt.assert_almost_equal(radiosity.patches_points.shape, (150, 4, 3))
-    npt.assert_almost_equal(radiosity.patches_area.shape, (150))
-    npt.assert_almost_equal(radiosity.patches_center.shape, (150, 3))
-    npt.assert_almost_equal(radiosity.patches_size.shape, (150, 3))
-    npt.assert_almost_equal(radiosity.patches_normal.shape, (150, 3))
+    radiosity = sp.DirectionalRadiosityFast.from_polygon(sample_walls, 0.5)
+    npt.assert_almost_equal(radiosity.patches_points.shape, (24, 4, 3))
+    npt.assert_almost_equal(radiosity.patches_area.shape, (24))
+    npt.assert_almost_equal(radiosity.patches_center.shape, (24, 3))
+    npt.assert_almost_equal(radiosity.patches_size.shape, (24, 3))
+    npt.assert_almost_equal(radiosity.patches_normal.shape, (24, 3))
 
 
 def test_check_visibility(sample_walls):
@@ -29,7 +29,7 @@ def test_check_visibility(sample_walls):
 
 
 def test_check_visibility_wrapper(sample_walls):
-    radiosity = sp.DirectionalRadiosityFast.from_polygon(sample_walls, 0.2)
+    radiosity = sp.DirectionalRadiosityFast.from_polygon(sample_walls, 0.5)
     radiosity.bake_geometry()
     visibility_matrix = sp.geometry._check_patch2patch_visibility(
         radiosity.patches_center, radiosity.patches_normal,
@@ -38,12 +38,9 @@ def test_check_visibility_wrapper(sample_walls):
 
 
 def test_compute_form_factors(sample_walls):
-    radiosity = sp.DirectionalRadiosityFast.from_polygon(sample_walls, 0.2)
+    radiosity = sp.DirectionalRadiosityFast.from_polygon(sample_walls, .5)
     radiosity.bake_geometry()
-    npt.assert_almost_equal(radiosity.form_factors.shape, (150, 150))
-    radiosity.bake_geometry()
-    npt.assert_almost_equal(radiosity.form_factors.shape, (150, 150))
-
+    npt.assert_almost_equal(radiosity.form_factors.shape, (24, 24))
 
 
 @pytest.mark.parametrize('walls', [