### Changes proposed in this pull request:

- complexity of tests is reduced, where ever possible
- example: on develop 126.07s and in this pr 96.74s on my local mashine


### 10 slowest tests now:
```
========================================================= slowest 10 durations ==========================================================
12.51s call     tests/test_DRadiosityFast.py::test_check_visibility
7.72s call     tests/test_multisource.py::test_multi_receiver[receivers0-frequencies0]
6.50s call     tests/test_DRadiosityFast_infinite_diffuse_plane.py::test_colocated_source_receiver[1]
5.30s call     tests/test_Radiosity.py::test_small_room_and_rotate
4.42s call     tests/test_Radiosity.py::test_small_room_and_shift
4.03s call     tests/test_multisource.py::test_multi_receiver[receivers0-frequencies1]
3.56s call     tests/test_DirectionalRadiosityFast.py::test_io[True-True]
2.96s call     tests/test_DRadiosityFast.py::test_init
2.01s call     tests/test_DRadiosityFast_order.py::test_order_vs_analytic[1]
1.82s call     tests/test_universal_formfactor.py::test_parallel_facing_patches[1.0-1.0-1.0]
============================================= 502 passed, 622 warnings in 96.74s (0:01:36) ==============================================
```

### 10 slowest test develop:
```
========================================================= slowest 10 durations ==========================================================
13.05s call     tests/test_Radiosity.py::test_small_room_and_rotate
11.91s call     tests/test_DRadiosityFast.py::test_check_visibility
10.07s call     tests/test_DirectionalRadiosityFast.py::test_io_within_simulation
8.04s call     tests/test_multisource.py::test_multi_receiver[receivers0-frequencies0]
6.30s call     tests/test_DRadiosityFast.py::test_compute_form_factors
6.16s call     tests/test_DRadiosityFast_infinite_diffuse_plane.py::test_colocated_source_receiver[1]
5.54s call     tests/test_DRadiosityFast.py::test_check_visibility_wrapper
4.79s call     tests/test_multisource.py::test_multi_receiver[receivers0-frequencies1]
4.37s call     tests/test_Radiosity.py::test_small_room_and_shift
4.07s call     tests/test_DRadiosityFast_order.py::test_order_vs_analytic[1]
============================================= 502 passed, 622 warnings in 126.07s (0:02:06) =============================================
```