@joao-fatela 2025-07-16
test_DRadiosityFast.py/test_check_visibility can also be optimized by increasing the patch size. Should I do it?